content-security-policy/frame-src/frame-src-5_1.html

<!DOCTYPE HTML>
<html>
<head>
    <title>iframe element src attribute must match src list</title>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
    <link rel='stylesheet' href='/resources/testharness.css'>
</head>
<body>
    <h1>iframe element src attribute must match src list</h1>
    <div id='log'></div>

<script>
var selfTest = async_test("Iframe loaded from self (as allowed by frame-src policy)");
var nonSelfTest = async_test("Iframe not loaded from www1.web-platform.test (as denied by frame-src policy)");
var brokenTest = async_test("blah");
setTimeout(function () {
    var ifr=document.getElementById('badiframe');
    ifr.onload=function(){
        brokenTest.step(function () {
            assert_false(true, "onload ran in bad iframe.");
        });
    };
    brokenTest.done();
}, 1000);

</script>

    <iframe src="frame-src-5_1yes.html" id="selfiFrame" onload='selfTest.done()'></iframe>
    <iframe src="http://www1.web-platform.test:8000/content-security-policy/frame-src/frame-src-5_1no.html" id="nonSelfiFrame" onload='nonSelfTest.step(function() {function() {test(assert_false(true, "loaded bad iframe from www1.web-platform.test (should have been denied by policy)"))})'></iframe>

    <iframe id="badiframe" src="/content-security-policy/frame-src/frame-src-5_1no.html" onload='brokenTest.step(function() {test(assert_false(true, "bad iframe"))})'></iframe>

<script>
nonSelfTest.done();
</script>


    <script async defer src='../support/checkReport.sub.js?reportField=violated-directive&reportValue=frame-src%20%27self%27'></script> 
</body>
</html>
